\name{ge.real.cts}
\alias{ge.real.cts}
\docType{data}
\title{General Electric Realized Objects}
\description{
  A list of General Electric transaction data aligned to calendar time sampling and already in realizedObjects.
}
\usage{data(ge.real.cts)}
\source{
http://faculty.washington.edu/ezivot/ezresearch.htm
}

\examples{
data(ge.real.cts)
plot(ge.real.cts[[1]]) 
}
\keyword{datasets}
