\name{ge.real.tts}
\alias{ge.real.tts}
\docType{data}
\title{General Electric Realized Objects}
\description{
  A list of General Electric transaction data aligned to tick time sampling and already in realizedObjects.
}
\usage{data(ge.real.tts)}
\source{
http://faculty.washington.edu/ezivot/ezresearch.htm
}

\examples{
data(ge.real.tts)
plot(ge.real.tts[[1]]) 
}
\keyword{datasets}
