\name{merge.realizedObject}
\alias{merge.realizedObject}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Merging realizedObjects}
\description{
  Merges realizedObjects, see users manual.
}
\usage{
merge.realizedObject(x, y = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ First object }
  \item{y}{ Second object }
  \item{...}{ Third, .... objects }
}
\details{
  see users manual.
}
\value{
  Multiple instrument realized object
}
\author{Scott Payseur <spayseur@u.washingon.edu>}

% R documentation directory.
\keyword{methods}