\name{msft.real.cts}
\alias{msft.real.cts}
\docType{data}
\title{Microsoft Realized Objects}
\description{
  A list of Microsoft transaction data aligned to calendar time sampling and already in realizedObjects.
}
\usage{data(msft.real.cts)}
\source{
http://faculty.washington.edu/ezivot/ezresearch.htm
}

\examples{
data(msft.real.cts)
plot(msft.real.cts[[1]]) 
}
\keyword{datasets}
