\name{msft.real.tts}
\alias{msft.real.tts}
\docType{data}
\title{Microsoft Realized Objects}
\description{
  A list of Microsoft transaction data aligned to tick time sampling and already in realizedObjects.
}
\usage{data(msft.real.tts)}
\source{
http://faculty.washington.edu/ezivot/ezresearch.htm
}

\examples{
data(msft.real.tts)
plot(msft.real.tts[[1]]) 
}
\keyword{datasets}
