\name{realized-internal}
\title{Internal realized objects}
\alias{.accum.naive}
\alias{.alignedAccum}
\alias{.alignReturns}
\alias{.convertData}
\alias{.corImage}
\alias{.First.lib}
\alias{.getDataCol}
\alias{.getDataColNum}
\alias{.kernel.chartoint}
\alias{.realized.variance}
\alias{[.realizedObject}
\alias{data.sameTime}
\alias{data.toCts}
\alias{data.toReturns}
\alias{dim.realizedObject}
\alias{getReturns}
\alias{plot.realizedObject}
\alias{print.realizedObject}
\alias{rRealized.variance}
\alias{ts2realized}
\alias{tsGetDay}
\alias{tsGetDayObject}
\alias{timeDate}
\description{Internal realized objects.}
\details{These are not to be called by the user.}
\keyword{internal}

