\name{realized-package}
\alias{realized-package}
\alias{realized}
\docType{package}
\title{
Realized Variance, Covariance and Correlation
}
\description{
Package contains various estimators and tools to research realized estimates of second moments.
}
\details{
\tabular{ll}{
Package: \tab realized\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2007-08-05\cr
License: \tab \cr
}
}
\author{
Scott Payseur <spayseur@u.washington.edu>
}
\references{
T.G. Andersen, T. Bollerslev, F.X. Diebold, and P. Labys. Great realizations. \emph{Risk}, 13:105 108, 2000.

Y. Fang. Volatility modeling and estimation of high-frequency data with gaussian noise. \emph{unpublished doctoral
thesis, MIT, Sloan School of Management}, 1996.

O. E. Barndorff-Nielson and N. Sheppard. Econometric analysis of realised volatility and its use in estimating
stochastic volatility models. \emph{Journal of the Royal Statistical Society}, Series B 64:253-280, 2002

T.G. Andersen, T. Bollerslev, F.X. Diebold, and P. Labys. The distribution of exchange rate volatility.
\emph{Journal of the American Statistical Association}, 96:42-55, 2001.

S. W. Payseur.  A One Day Comparison of Realized Variance and Covariance Estimators. \emph{Working Paper: University 
of Washington}, 2007

L. Zhang, P.A Mykland, and Y. Ait-Sahalia. A tale of two time scales: Determining integrated volatility
with noisy high-frequency data. \emph{Journal of the American Statistical Association}, 2005.

Michiel de Pooter, Martin Martens, and Dick van Dijk. Predicting the daily covariance matrix for sp100
stocks using intraday data - but which frequency to use? \emph{Working Paper}, October 2005.

Ole E. Barndorff-Nielsen, Peter Reinhard Hansen, Asger Lunde, and Neil Shephard. Regular and modified kernel-based estimators of integrated variance: The case with independent noise. \emph{Working Paper}, 2004.

J. E. Griffen and R. C. A. Oomen. Covariance measurement in the presence of non-synchronous trading and
market microstructure noise. \emph{Working Paper}, June 27th, 2006.

T.G. Andersen, T. Bollerslev, F.X. Diebold, and P. Labys. The distribution of exchange rate volatility.
\emph{Journal of the American Statistical Association}, 96:42-55, 2001.



}

\keyword{ package }

\examples{
# see users manual
}