\name{rKernel.available}
\alias{rKernel.available}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Available Kernels}
\description{
Returns a vector of the available kernels.
}
\usage{
rKernel.available()
}
\references{
Ole E. Barndorff-Nielsen, Peter Reinhard Hansen, Asger Lunde, and Neil Shephard. Regular and modified kernel-based estimators of integrated variance: The case with independent noise. \emph{Working Paper}, 2004.
 
 
 }

\author{ Scott Payseur <spayseur@u.washington.edu>}

\seealso{\code{\link{rc.kernel}}, \code{\link{rv.kernel}}, \code{\link{rKernel}}, \code{\link{rRealizedVariance}} }
\examples{

rKernel.available()

}
\keyword{methods}
